Adding TDSReportValidator interface, breaking out impl class
